requirements.txt
ansicon==1.89.0
arabic-reshaper==2.1.4
arrow==1.2.3
asgiref==3.5.2
asn1crypto==1.5.1
blessed==1.19.1
Brotli==1.0.9
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
contourpy==1.0.6
creds==1.0.15
cryptography==38.0.1
cssselect2==0.7.0
cycler==0.11.0
Django==4.1.1
django-bootstrap-datepicker-plus==5.0.2
django-crispy-forms==1.14.0
django-extensions==3.2.1
django-filter==22.1
django-htmx==1.12.2
django-mathfilters==1.0.0
django-picklefield==3.1
django-q==1.3.9
django-tables2==2.4.1
django-tinymce==3.5.0
docraptor==3.0.0
et-xmlfile==1.1.0
fonttools==4.38.0
future==0.18.2
html5lib==1.1
idna==3.4
Jinja2==3.1.2
jinxed==1.2.0
kiwisolver==1.4.4
lxml==4.9.1
MarkupSafe==2.1.1
matplotlib==3.6.2
numpy==1.23.4
openpyxl==3.0.10
oscrypto==1.3.0
packaging==21.3
pandas==1.5.0
pdfcrowd==5.9.0
pdfkit==1.0.0
Pillow==9.2.0
pyactiveresource==2.2.2
pyclean==2.2.0
pycparser==2.21
pydantic==1.10.2
pydyf==0.5.0
pyHanko==0.15.0
pyhanko-certvalidator==0.19.5
PyJWT==2.5.0
pyparsing==3.0.9
PyPDF3==1.0.6
pyphen==0.13.0
python-bidi==0.4.2
python-dateutil==2.8.2
pytz==2022.4
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.3.1
redis==3.5.3
reportlab==3.6.11
requests==2.28.1
seaborn==0.12.1
ShopifyAPI==12.0.0
six==1.16.0
sqlparse==0.4.2
svglib==1.4.1
tablib==3.2.1
tinycss2==1.2.1
tqdm==4.64.1
typing_extensions==4.4.0
tzdata==2022.2
tzlocal==4.2
uritools==4.0.0
urllib3==1.26.12
wcwidth==0.2.5
weasyprint==57.0
webencodings==0.5.1
xhtml2pdf==0.2.8
xlwt==1.3.0
zopfli==0.2.1
